Rename Database/Exception to Database/Error.
